apps/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.27.0
python-dotenv==1.0.1
sqlalchemy==2.0.36
pydantic==2.10.4
